+++ ios_example/mid-example/MidExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  MidExampleApp.swift
+//  mid-example
+//
+//  Created by doroved on 21.01.2026.
+//
+
+import SwiftUI
+
+@main
+struct MidExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}