+++ src/android.rs
@@ -1,0 +1,4 @@
+// #[cfg(target_os = "android")]
+// use crate::errors::MIDError;
+
+// Is implementation for Android needed?